% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_url}
\alias{check_url}
\title{Check a single URL}
\usage{
check_url(url, options = lychee_options())
}
\arguments{
\item{url}{A single URL string to check.}

\item{options}{A \code{lychee_options} object created by \code{\link[=lychee_options]{lychee_options()}}.}
}
\value{
A list with \code{url}, \code{is_success}, \code{code}, and \code{details}.
}
\description{
Check a single URL
}
\examples{
\dontrun{
check_url("https://www.r-project.org")
}
}
