% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{check_url}
\alias{check_url}
\title{Check a single URL with lychee and return its status.}
\usage{
check_url(url)
}
\arguments{
\item{url}{A single URL string to check.}
}
\value{
A list with \code{url}, \code{is_success}, \code{code}, and \code{details}.
}
\description{
Check a single URL with lychee and return its status.
}
