% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rambutan-package.R
\docType{package}
\name{rambutan-package}
\alias{rambutan}
\alias{rambutan-package}
\title{rambutan: Fast Link Checking Powered by 'lychee'}
\description{
An interface to the 'lychee' Rust crate
(\url{https://github.com/lycheeverse/lychee}), a fast, async link checker.
Wraps 'lychee-lib' via 'extendr' to check URLs, files, and directories
for broken links.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://drmowinckels.io/rambutan/}
  \item Report bugs at \url{https://github.com/drmowinckels/rambutan/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID}) (drmowinckels)

}
\keyword{internal}
